Extract createSongs helper
